# Conflicts:
#	lib/SQLite/database_helper.dart
#	lib/Views/profile.dart
#	pubspec.lock
#	pubspec.yaml

+++ lib/Views/profile.dart
@@ -8,21 +8,57 @@ import 'auth.dart';
 import 'changePassword.dart';
 import 'change_email.dart';
 import 'user_screen_list.dart';
+import 'user_status_list.dart'; // Nueva importación para la lista de estado de usuarios
 import 'theme_settings_page.dart';
 
-class Profile extends StatelessWidget {
+class Profile extends StatefulWidget {
   final Users? profile;
 
   const Profile({super.key, this.profile});
 
   @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
+  State<Profile> createState() => _ProfileState();
+}
 
+class _ProfileState extends State<Profile> {
+  bool isActive = false; // Estado inicial del usuario
+
+  @override
+  void initState() {
+    super.initState();
+    fetchUserStatus(); // Obtener estado de conexión inicial
+  }
+
+  Future<void> fetchUserStatus() async {
+    final db = DatabaseHelper();
+    Users? user = await db.getUser(widget.profile!.usrName);
+    if (user != null) {
+      setState(() {
+        isActive = user.isActive;
+      });
+    }
+  }
+
+  Future<void> toggleUserStatus(bool status) async {
+    final db = DatabaseHelper();
+    await db.updateConnectionStatus(widget.profile!.usrName, status);
+    setState(() {
+      isActive = status;
+    });
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(
+          status ? "Estás ahora Activo." : "Estás ahora Desconectado.",
+        ),
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: const Text("Gestión de Perfil"),
-        backgroundColor: theme.appBarTheme.backgroundColor,
         actions: [
           IconButton(
             icon: const Icon(Icons.help_outline),
@@ -53,33 +89,38 @@ class Profile extends StatelessWidget {
           child: Padding(
             padding: const EdgeInsets.symmetric(vertical: 45.0, horizontal: 20),
             child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
               children: [
                 CircleAvatar(
-                  backgroundColor: theme.primaryColor,
-                  radius: 77,
-                  child: const CircleAvatar(
-                    backgroundImage: AssetImage("assets/no user.png"),
-                    radius: 75,
-                  ),
+                  backgroundColor: Theme.of(context).primaryColor,
+                  radius: 75,
+                  backgroundImage: const AssetImage("assets/no user.png"),
                 ),
                 const SizedBox(height: 10),
                 Text(
-                  "Bienvenido ${profile?.fullName}",
+                  "Bienvenido ${widget.profile?.fullName}",
                   style: TextStyle(
-                    color: theme.primaryColor,
-                    fontSize: 30,
+                    fontSize: 25,
                     fontWeight: FontWeight.bold,
+                    color: Theme.of(context).primaryColor,
                   ),
                 ),
-                Text(
-                  "${profile?.email}",
-                  style: const TextStyle(
-                    color: Colors.black45,
-                    fontSize: 15,
-                    fontWeight: FontWeight.bold,
-                  ),
+                const SizedBox(height: 10),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    const Text(
+                      "Estado:",
+                      style: TextStyle(fontSize: 18),
+                    ),
+                    Switch(
+                      value: isActive,
+                      onChanged: toggleUserStatus,
+                      activeColor: Colors.green,
+                      inactiveThumbColor: Colors.red,
+                    ),
+                  ],
                 ),
+                const SizedBox(height: 20),
                 Button(
                   label: "Logout",
                   press: () {
@@ -89,59 +130,70 @@ class Profile extends StatelessWidget {
                     );
                   },
                 ),
-
                 ListTile(
-                  leading: Icon(Icons.person, color: theme.primaryColor, size: 30),
+                  leading: Icon(Icons.person, size: 30, color: Theme.of(context).primaryColor),
                   subtitle: const Text("Nombre completo"),
-                  title: Text("${profile?.fullName}"),
+                  title: Text("${widget.profile?.fullName}"),
                 ),
                 ListTile(
-                  leading: Icon(Icons.account_circle, color: theme.primaryColor, size: 30),
+                  leading: Icon(Icons.account_circle, size: 30, color: Theme.of(context).primaryColor),
                   subtitle: const Text("Usuario"),
-                  title: Text("${profile?.usrName}"),
+                  title: Text("${widget.profile?.usrName}"),
                 ),
                 ListTile(
-                  leading: Icon(Icons.lock, color: theme.primaryColor, size: 30),
+                  leading: Icon(Icons.lock, size: 30, color: Theme.of(context).primaryColor),
                   subtitle: const Text("Seguridad"),
                   title: const Text("Cambiar Contraseña"),
                   onTap: () {
                     Navigator.push(
                       context,
                       MaterialPageRoute(
-                        builder: (context) => ChangePasswordScreen(profile: profile!),
+                        builder: (context) => ChangePasswordScreen(profile: widget.profile!),
                       ),
                     );
                   },
                 ),
                 ListTile(
-                  leading: Icon(Icons.email, color: theme.primaryColor, size: 30),
-                  subtitle: const Text("Cambiar Correo Electrónico"),
-                  title: Text("${profile?.email}"),
+                  leading: Icon(Icons.email, size: 30, color: Theme.of(context).primaryColor),
+                  subtitle: const Text("Correo Electrónico"),
+                  title: Text("${widget.profile?.email}"),
                   onTap: () {
                     Navigator.push(
                       context,
                       MaterialPageRoute(
-                        builder: (context) => ChangeEmailScreen(profile: profile!),
+                        builder: (context) => ChangeEmailScreen(profile: widget.profile!),
                       ),
                     );
                   },
                 ),
                 ListTile(
-                  leading: Icon(Icons.list, color: theme.primaryColor, size: 30),
-                  subtitle: const Text("Gestionar usuarios"),
-                  title: const Text("Lista de Usuarios"),
+                  leading: Icon(Icons.list, size: 30, color: Theme.of(context).primaryColor),
+                  subtitle: const Text("Bloqueo de Usuarios"),
+                  title: const Text("Lista de Bloqueo"),
                   onTap: () {
                     Navigator.push(
                       context,
                       MaterialPageRoute(
-                        builder: (context) => UserListScreen(
-                          currentUserName: profile?.usrName ?? "",
-                        ),
+                        builder: (context) =>
+                            UserListScreen(currentUserName: widget.profile?.usrName ?? ""),
                       ),
                     );
                   },
                 ),
-                SizedBox(height: 20), // Espaciado
+                ListTile(
+                  leading: Icon(Icons.network_check, size: 30, color: Theme.of(context).primaryColor),
+                  subtitle: const Text("Estado de Usuarios"),
+                  title: const Text("Lista de Estado"),
+                  onTap: () {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => const UserStatusListScreen(),
+                      ),
+                    );
+                  },
+                ),
+                const SizedBox(height: 20), // Espaciado
                 Button(
                   label: "Eliminar Cuenta",
                   press: () => _showDeleteConfirmationDialog(context),
@@ -153,47 +205,52 @@ class Profile extends StatelessWidget {
       ),
     );
   }
-  // Metodo para mostrar el cuadro de diálogo de confirmación
+
   void _showDeleteConfirmationDialog(BuildContext context) {
     showDialog(
-        context: context,
-        builder: (BuildContext context) {
-      return AlertDialog(
-          title: Text("Confirmar Eliminación"),
-          content: Text(
-            "¿Estás seguro de que quieres eliminar tu cuenta? Esta acción es irreversible y toda tu información será eliminada.",
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: const Text("Confirmar Eliminación"),
+          content: const Text(
+            "¿Estás seguro de que quieres eliminar tu cuenta? Esta acción es irreversible.",
           ),
           actions: [
-          TextButton(
-          onPressed: () {
-        Navigator.pop(context); // Cierra el cuadro de diálogo
-      },
-    child: Text("Cancelar", style: TextStyle(color: Colors.grey)),
-    ),
+            TextButton(
+              onPressed: () {
+                Navigator.pop(context);
+              },
+              child: const Text(
+                "Cancelar",
+                style: TextStyle(color: Colors.grey),
+              ),
+            ),
             TextButton(
               onPressed: () async {
-                Navigator.pop(context); // Cierra el cuadro de diálogo primero
+                Navigator.pop(context); // Cierra el diálogo primero
                 await _deleteUserAccount(context); // Luego procede con la eliminación
               },
-              child: Text("Eliminar", style: TextStyle(color: Colors.red)),
+              child: const Text(
+                "Eliminar",
+                style: TextStyle(color: Colors.red),
+              ),
             ),
           ],
-      );
-        },
+        );
+      },
     );
   }
-// Metodo para eliminar el usuario
+
   Future<void> _deleteUserAccount(BuildContext context) async {
     final db = DatabaseHelper();
 
     try {
-      // Asegúrate de que el usuario esté disponible
-      final String currentUserEmail = profile!.email!;
+      final String currentUserEmail = widget.profile!.email!;
       bool success = await db.deleteUserAccount(currentUserEmail);
 
       if (success) {
         ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text("Cuenta eliminada con éxito.")),
+          const SnackBar(content: Text("Cuenta eliminada con éxito.")),
         );
 
         // Redirige al inicio de sesión
@@ -204,13 +261,13 @@ class Profile extends StatelessWidget {
         );
       } else {
         ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text("Error al eliminar la cuenta. Inténtalo de nuevo.")),
+          const SnackBar(content: Text("Error al eliminar la cuenta. Inténtalo de nuevo.")),
         );
       }
     } catch (e) {
       print("Error al eliminar la cuenta: $e");
       ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text("Ocurrió un error inesperado. Por favor, intenta más tarde.")),
+        const SnackBar(content: Text("Ocurrió un error inesperado. Inténtalo más tarde.")),
       );
     }
   }
